cleanup, reformating
