+++ fcode.sh
@@ -1,0 +1,6 @@
+# format code
+isort src/microcirculation
+black src/microcirculation
+
+isort tests
+black tests
