+++ src/microcirculation/filters/standard_transformations.py
@@ -1,18 +1,19 @@
-import numpy as np
+from typing import Iterable
+
 import cv2
+import numpy as np
 from PIL import Image
 from PIL.ImageFilter import GaussianBlur
-from typing import Iterable
 
 __all__ = [
     "gaussian_filter",
     "gamma_transform",
     "threshold",
-    "global_histogram_equalization",
-    "local_histogram_equalization",
+    "histogram_equalization_global",
+    "histogram_equalization_local",
     "median_blur",
     "detect_edges_sobel",
-    "normalize_frames_brightness"
+    "normalize_frames_brightness",
 ]
 
 
@@ -28,25 +29,15 @@ def gamma_transform(image: Image, gamma: float = 0.5) -> Image:
     return Image.fromarray(image)
 
 
-def global_histogram_equalization(image: Image) -> Image:
-    image_matrix = np.array(image)
-    return Image.fromarray(cv2.equalizeHist(image_matrix))
-
-
-def local_histogram_equalization(image: Image) -> Image:
-    """Apply local histogram equalization using CLAHE"""
-    image_matrix = np.array(image)
-    clahe = cv2.createCLAHE(clipLimit=2.0, tileGridSize=(8,8))
-    return Image.fromarray(clahe.apply(image_matrix))
-    
-
-# def box_filter(image: Image) -> Image:
-#     image_matrix = np.array(image)
-
-
 def median_blur(image: Image) -> Image:
-    image_matrix = np.array(image)
-    return Image.fromarray(cv2.medianBlur(image_matrix, 3))
+    """Apply median blur.
+
+    The function smoothes an image using the median filter with the ksize×ksize aperture.
+    Each channel of a multi-channel image is processed independently. In-place operation is supported.
+    """
+    frame = np.array(image)
+    frame_blur = cv2.medianBlur(frame, ksize=3)
+    return Image.fromarray(frame_blur)
 
 
 def threshold(image: Image, value: int, invert: bool = False) -> Image:
@@ -87,11 +78,41 @@ def detect_edges_sobel(image: Image) -> Image:
 
     return Image.fromarray(edges)
 
-def normalize_frames_brightness(frames: Iterable[np.array]):
+
+def histogram_equalization_global(image: Image) -> Image:
+    """Global histogram equalization of single frame.
+
+    Requires greyscale imge.
+    It is a method that improves the contrast in an image, in order to stretch out the intensity range
+    """
+    frame = np.array(image)
+    frame_equalized = cv2.equalizeHist(frame)
+    return Image.fromarray(frame_equalized)
+
+
+def histogram_equalization_local(image: Image) -> Image:
+    """Apply local histogram equalization using CLAHE.
+
+    https://docs.opencv.org/3.4/d6/db6/classcv_1_1CLAHE.html
+
+        clipLimit: This is the threshold for contrast limiting
+        tileGridSize: Divides the input image into M x N tiles and then applies histogram equalization
+        to each local tile
+    """
+
+    frame = np.array(image)
+    clahe = cv2.createCLAHE(clipLimit=2.0, tileGridSize=(1, 1))
+    return Image.fromarray(clahe.apply(frame))
+
+
+def normalize_frames_brightness(frames: Iterable[np.array]) -> Iterable[np.array]:
+    """Normalizes the brightness of multiple frames.
+
+    FIXME: check this
+    """
     max_intensity = np.max(frames)
     avg_intensity = np.mean(frames)
-    frames = frames/max_intensity
-    frames = (frames*avg_intensity)
+    frames = frames / max_intensity
+    frames = frames * avg_intensity
 
-    # print(frames)
     return frames
