+++ src/microcirculation/utils.py
@@ -1,7 +1,7 @@
-from PIL import Image
-import numpy as np
 from typing import Iterable
-from video_info import get_keypoints_for_frame
+
+import numpy as np
+from PIL import Image
 
 
 def get_average_grayscale_value(image: Image.Image) -> int:
@@ -24,9 +24,10 @@ def stack_images(images: Iterable[Image.Image]) -> Image.Image:
 
     return stacked_image
 
+
 def get_image_segment(
-    image: Image, 
-    left_perc: int, 
+    image: Image,
+    left_perc: int,
     right_perc: int,
     top_perc: int,
     bottom_perc: int,
@@ -42,11 +43,10 @@ def get_image_segment(
     """
 
     (width, height) = image.size
-    region = (left_perc/100 * width, top_perc/100 * height, right_perc/100 * width, bottom_perc/100 * height)
+    region = (
+        left_perc / 100 * width,
+        top_perc / 100 * height,
+        right_perc / 100 * width,
+        bottom_perc / 100 * height,
+    )
     return image.crop(region)
-
-
-def keypoint_detection(frame: Image, kp_method: str):
-    """Keypoint detection"""
-    frame = np.array(frame)
-    return get_keypoints_for_frame(frame, kp_method)
