+++ src/microcirculation/video/conversion.py
@@ -1,7 +1,6 @@
 from pathlib import Path
 from typing import Tuple
 
-import numpy as np
 import cv2 as cv
